+++ movie/infra/src/main/kotlin/com/example/movie/persistence/genre/model/GenreEntity.kt
@@ -1,4 +1,4 @@
-package com.example.movie.entity
+package com.example.movie.persistence.genre.model
 
 import com.example.movie.domain.movie.model.Genre
 import com.example.movie.persistence.common.BaseEntity
