+++ movie/infra/src/main/kotlin/com/example/movie/persistence/genre/repository/GenreJpaRepository.kt
@@ -1,6 +1,6 @@
-package com.example.movie.repository
+package com.example.movie.persistence.genre.repository
 
-import com.example.movie.entity.GenreEntity
+import com.example.movie.persistence.genre.model.GenreEntity
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
